requirements.txt
edge_tts==7.2.7
openai==2.13.0
openai_whisper==20250625
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.52.1
torch==2.9.1
